Maybe finished, doesn't work though

+++ src/plane.cpp
@@ -13,9 +13,9 @@ Plane::Plane(int verticesPerEdge)
 
 	std::vector<float> vertices(vertexCount * floatsPerVertex);
 
-	float stepSize{ 1.0 / mVerticesPerEdge - 1 };
-	float currentXPos{ -0.5 };
-	float currentZPos{ -0.5 };
+	double stepSize{ 1.0 / mVerticesPerEdge - 1 };
+	double currentXPos{ -0.5 };
+	double currentZPos{ -0.5 };
 	for (int vertexI{ 0 }; vertexI < vertexCount; ++vertexI) {
 		vertices[vertexI * floatsPerVertex + 0] = currentXPos;
 		vertices[vertexI * floatsPerVertex + 1] = currentZPos;
@@ -60,9 +60,9 @@ void Plane::rebuild(int verticesPerEdge) {
 
 	std::vector<float> vertices(vertexCount * floatsPerVertex);
 
-	float stepSize{ 1.0 / mVerticesPerEdge - 1) };
-	float currentXPos{ -0.5 };
-	float currentZPos{ -0.5 };
+	double stepSize{ 1.0 / (mVerticesPerEdge - 1) };
+	double currentXPos{ -0.5 };
+	double currentZPos{ -0.5 };
 	for (int vertexI{ 0 }; vertexI < vertexCount; ++vertexI) {
 		vertices[vertexI * floatsPerVertex + 0] = currentXPos;
 		vertices[vertexI * floatsPerVertex + 1] = currentZPos;
